split forms to forders
