+++ src/components/forms/formsLogo/formLogo.styles.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+export const Title = styled.div`
+  display: flex;
+  margin-bottom: 50px;
+`;
+
+export const TitleText = styled.h1`
+  font-weight: 700;
+  font-size: 30px;
+  line-height: 45px;
+`;
+
+export const Logo = styled.div`
+  margin-right: 20px;
+`;
