+++ src/components/forms/loginForm/LoginForm.styles.js
@@ -7,23 +7,10 @@ export const Container = styled.div`
   align-items: center;
 `;
 
-export const Title = styled.div`
-  display: flex;
-  margin-bottom: 50px;
-`;
-
-export const TitleText = styled.h1`
-  font-weight: 700;
-  font-size: 30px;
-  line-height: 45px;
-`;
-
-export const Logo = styled.div`
-  margin-right: 20px;
-`;
-
 export const FormikForm = styled(Form)`
   display: flex;
   flex-direction: column;
   justify-content: center;
+  align-items: center;
+  width: 410px;
 `;
